+++ web/src/LandingPage.js
@@ -3,7 +3,7 @@ import styled, { css } from 'react-emotion'
 import { NavLink } from 'react-router-dom'
 import { H1, H2, H3, theme, mediaQuery } from './styles'
 import Layout from './components/Layout'
-import { buttonStyles } from './forms/Button'
+import { buttonStyles } from './components/forms/Button'
 import { Trans } from 'lingui-react'
 import { Helmet } from 'react-helmet'
