+++ web/src/RegistrationPage.js
@@ -4,10 +4,10 @@ import { Trans } from 'lingui-react'
 import { css } from 'react-emotion'
 import { theme, visuallyhidden, mediaQuery } from './styles'
 import Layout from './components/Layout'
-import { TextFieldAdapter, TextAreaAdapter } from './forms/TextInput'
-import FieldSet from './forms/FieldSet'
-import { RadioAdapter } from './forms/MultipleChoice'
-import Button from './forms/Button'
+import { TextFieldAdapter, TextAreaAdapter } from './components/forms/TextInput'
+import FieldSet from './components/forms/FieldSet'
+import { RadioAdapter } from './components/forms/MultipleChoice'
+import Button from './components/forms/Button'
 import { Form, Field } from 'react-final-form'
 import { withApollo } from 'react-apollo'
 import gql from 'graphql-tag'
