+++ web/src/ReviewPage.js
@@ -6,7 +6,7 @@ import { NavLink, Redirect } from 'react-router-dom'
 import { H1, theme, BottomContainer, TopContainer } from './styles'
 import Layout from './components/Layout'
 import { GET_USER_DATA, SUBMIT } from './queries'
-import Button from './forms/Button'
+import Button from './components/forms/Button'
 import Summary from './components/Summary'
 import { Submission } from './components/Submission'
 import { Reminder } from './components/Reminder'
