+++ web/src/components/forms/FieldSet.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { css } from 'react-emotion'
-import { theme, visuallyhidden, mediaQuery } from '../styles'
+import { theme, visuallyhidden, mediaQuery } from '../../styles'
 
 const fieldset = css`
   border: none;
